Refactor the talk template and eliminate the workshop template
