+++ gatsby-node.js
@@ -12,7 +12,7 @@ exports.createPages = ({ actions, graphql }) => {
   const templates = {
     talks: path.resolve("src/templates/talk.js"),
     sessions: path.resolve("src/templates/session.js"),
-    workshops: path.resolve("src/templates/workshop.js"),
+    workshops: path.resolve("src/templates/talk.js"),
   }
 
   return graphql(`
