+++ src/components/talk/participate.js
@@ -1,0 +1,12 @@
+import React from "react"
+import { Row } from "reactstrap"
+
+const Participate = ({ href }) => {
+  return (
+    <Row>
+      <a href={href}>Participa</a>
+    </Row>
+  )
+}
+
+export default Participate
