+++ src/components/talk/speaker.js
@@ -1,6 +1,6 @@
 import React from "react"
 import Img from "gatsby-image"
-import Occupations from "./occupations"
+import Occupations from "../occupations"
 
 const styles = {
   image: {
