+++ src/data/events/talks/test1.md
@@ -1,0 +1,25 @@
+---
+title: Software Engineering at Outsystems
+
+speakers:
+  - name: Ricardo Araújo
+    occupations:
+      - what: senior software engineer
+        where: outsystems
+    bio: Almadense de gema, Proencense de coração, Ricardo Araújo tem 20 anos de carreira ligado ao desenvolvimento web, tendo aprendido e desenvolvido aplicações para diversas indústrias e clientes de diversos países. Em 2009 abriu o centro de desenvolvimento remoto da OutSystems situado em Proença-a-Nova e de lá tem ajudado a OutSystems a liderar o mercado de Low-Code. O actual desafio é o de aumentar o ecosistema em 50 000 programadores até final de 2020.
+
+    img: "../../../images/JoseCarreiras.jpg"
+    linkedin: https://www.linkedin.com/in/ricardonevesaraujo/
+    website: https://davidsopas.com
+    twitter: https://twitter.com/dsopas
+
+day: 2019-10-29
+start_time: 15h00
+end_time: 16h00
+place: B032
+type: 0-talk
+
+path: /talks/software-engineering-at-outsystems
+---
+
+Todos os procuramos do que se faz bem lá fora esquecendo que em Portugal há equipas de classe mundial a inovar nos processos de engenharia de software. É isso que pretendemos dar a conhecer durante a nossa apresentação, quais os ingredientes essenciais para uma empresa de software nascida e baseada em Portugal
